# backend/mja-api/src/main/resources/META-INF/openapi/openapi.yaml
...
openapi: 3.0.3
info:
  title: mja-api
  description: API für mein Internet-Aufgabenarchiv
  contact:
    name: Heike Winkelvoß
    url: https://mathe-jung-alt.de
    email: info@egladil.de
  license:
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.4.0
servers:
- url: http://localhost:9210
  description: Auto generated value
- url: http://0.0.0.0:9210
  description: Auto generated value
...
paths:
  /deskriptoren/ids:
    get:
      tags:
      - Deskriptoren
      summary: Wandelt Deskriptoren in eine kommaseparierte Liste ihrer IDs um
      operationId: transformToDeskriptorenOrdinal
      parameters:
      - name: deskriptoren
        in: query
        description: kommaseparierte Liste von Namen von Deskriptoren
        schema:
          pattern: "^[a-zA-ZäöüßÄÖÜ\\d\\,\\- ]{0,200}$"
          type: string
      responses:
        "200":
          description: Deskriptoren erfolgreich transformiert.
          content:
            text/plain:
              schema:
                type: string
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      security:
      - SecurityScheme:
        - ADMIN
        - AUTOR
      - SecurityScheme:
        - ADMIN
        - AUTOR
...
  /raetsel/admin/v2:
    get:
      tags:
      - Raetsel
      summary: "Gibt alle Rätsel zurück, die auf die gegebene Suchanfrage passen."
      operationId: findRaetsel
      parameters:
      - name: suchstring
        in: query
        description: "Freitext zum suchen. Es erfolgt eine Volltextsuche über Schlü\
          ssel, Name, Kommentar, Frage und Lösung"
        schema:
          pattern: "^[\\w äöüß \\+ \\- \\. \\,]{4,30}$"
          type: string
      - name: deskriptoren
        in: query
        description: kommaseparierte Liste von Deskriptoren-Identifizierern
        schema:
          pattern: "^[a-zA-ZäöüßÄÖÜ\\d\\,\\- ]{0,200}$"
          type: string
      - name: typeDeskriptoren
        in: query
        description: wie die Deskriptoren gesendet werden (NAME oder ID)
        required: true
        schema:
          $ref: '#/components/schemas/EnumType'
      - name: limit
        in: query
        description: "Pagination: pageSize"
        schema:
          format: int32
          default: 20
          type: integer
      - name: offset
        in: query
        description: "Pagination: pageIndex"
        schema:
          format: int32
          default: 0
          type: integer
      - name: sortDirection
        in: query
        description: Sortierung. Es wird nach SCHLUESSEL sortiert.
        schema:
          allOf:
          - $ref: '#/components/schemas/SortDirection'
          - default: asc
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RaetselsucheTreffer'
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      security:
      - SecurityScheme:
        - ADMIN
        - AUTOR
      - SecurityScheme:
        - ADMIN
        - AUTOR
  /raetsel/v1:
    get:
      tags:
      - Raetsel
      summary: "Gibt alle Rätsel zurück, die auf die gegebene Suchanfrage passen.\
        \ Wird mit dem nächsten PROD-Release gestrichen."
      operationId: findRaetsel
      parameters:
      - name: suchstring
        in: query
        description: "Freitext zum suchen. Es erfolgt eine Volltextsuche über Schlü\
          ssel, Name, Kommentar, Frage und Lösung"
        schema:
          pattern: "^[\\w äöüß \\+ \\- \\. \\,]{4,30}$"
          type: string
      - name: deskriptoren
        in: query
        description: kommaseparierte Liste von Deskriptoren-Identifizierern
        schema:
          pattern: "^[a-zA-ZäöüßÄÖÜ\\d\\,\\- ]{0,200}$"
          type: string
      - name: typeDeskriptoren
        in: query
        description: wie die Deskriptoren gesendet werden (NAME oder ID)
        required: true
        schema:
          $ref: '#/components/schemas/EnumType'
      - name: limit
        in: query
        description: "Pagination: pageSize"
        schema:
          format: int32
          default: 20
          type: integer
      - name: offset
        in: query
        description: "Pagination: pageIndex"
        schema:
          format: int32
          default: 0
          type: integer
      - name: sortDirection
        in: query
        description: Sortierung. Es wird nach SCHLUESSEL sortiert.
        schema:
          allOf:
          - $ref: '#/components/schemas/SortDirection'
          - default: asc
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RaetselsucheTreffer'
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      deprecated: true
      security:
      - SecurityScheme:
        - ADMIN
        - AUTOR
      - SecurityScheme:
        - ADMIN
        - AUTOR
    put:
      tags:
      - Raetsel
      summary: vorhandenes Rätsel ändern
      operationId: raetselAendern
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EditRaetselPayload'
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Raetsel'
        "409":
          description: der geänderte schluessel ist schon vergeben
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessagePayload'
        "500":
          description: server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessagePayload'
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      security:
      - SecurityScheme:
        - ADMIN
        - AUTOR
      - SecurityScheme:
        - ADMIN
        - AUTOR
    post:
      tags:
      - Raetsel
      summary: neues Rätsel anlegen
      operationId: raetselAnlegen
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EditRaetselPayload'
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Raetsel'
        "409":
          description: der gewählte schluessel ist schon vergeben
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessagePayload'
        "500":
          description: server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessagePayload'
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      security:
      - SecurityScheme:
        - ADMIN
        - AUTOR
      - SecurityScheme:
        - ADMIN
        - AUTOR
  /raetsel/v1/PDF/{raetselID}:
    get:
      tags:
      - Raetsel
      summary: generiert ein PDF mit dem Rätsel
      operationId: raetselPDFGenerieren
      parameters:
      - name: raetselID
        in: path
        description: technische ID des Rätsels
        required: true
        schema:
          pattern: "^[a-fA-F\\d\\-]{1,36}$"
          type: string
      - name: layoutAntwortvorschlaege
        in: query
        description: "Layout, wie die Antwortvorschläge dargestellt werden sollen,\
          \ wenn es welche gibt (Details siehe LayoutAntwortvorschlaege)"
        required: true
        schema:
          $ref: '#/components/schemas/LayoutAntwortvorschlaege'
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeneratedFile'
        "500":
          description: server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessagePayload'
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      security:
      - SecurityScheme:
        - ADMIN
        - AUTOR
        - LEHRER
        - PRIVAT
        - STANDARD
      - SecurityScheme:
        - ADMIN
        - AUTOR
        - LEHRER
        - PRIVAT
        - STANDARD
  /raetsel/v1/PNG/{raetselID}:
    post:
      tags:
      - Raetsel
      summary: generiert die Vorschaubilder (png) des Rätsels
      operationId: raetselImagesGenerieren
      parameters:
      - name: raetselID
        in: path
        required: true
        schema:
          type: string
      - name: layoutAntwortvorschlaege
        in: query
        required: true
        schema:
          $ref: '#/components/schemas/LayoutAntwortvorschlaege'
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Images'
        "500":
          description: server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessagePayload'
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      security:
      - SecurityScheme:
        - ADMIN
        - AUTOR
      - SecurityScheme:
        - ADMIN
        - AUTOR
  /raetsel/v1/PNG/{schluessel}:
    get:
      tags:
      - Raetsel
      summary: Läd die Vorschaubilder (png) des Rätsels
      operationId: raetselImagesLaden
      parameters:
      - name: schluessel
        in: path
        description: Fachlicher Schlüssel des Rätsels
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Images'
        "500":
          description: server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessagePayload'
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      security:
      - SecurityScheme:
        - ADMIN
        - AUTOR
      - SecurityScheme:
        - ADMIN
        - AUTOR
  /raetsel/v1/{raetselID}:
    get:
      tags:
      - Raetsel
      summary: Läd die Details des Rätsels mit der gegebenen ID
      operationId: raetselDetailsLaden
      parameters:
      - name: raetselID
        in: path
        description: technische ID des Rätsels
        required: true
        schema:
          pattern: "^[a-fA-F\\d\\-]{1,36}$"
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Raetsel'
        "404":
          description: Gibt es nicht
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessagePayload'
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      security:
      - SecurityScheme:
        - ADMIN
        - AUTOR
      - SecurityScheme:
        - ADMIN
        - AUTOR
  /raetsel/v2:
    get:
      tags:
      - Raetsel
      summary: Gibt alle Rätsel mit den gegebenen Deskriptoren zurück
      operationId: findRaetselPublic
      parameters:
      - name: deskriptoren
        in: query
        description: kommaseparierte Liste von Deskriptoren-Identifizierern
        schema:
          pattern: "^[a-zA-ZäöüßÄÖÜ\\d\\,\\- ]{0,200}$"
          type: string
      - name: typeDeskriptoren
        in: query
        description: wie die Deskriptoren gesendet werden (NAME oder ID)
        required: true
        schema:
          $ref: '#/components/schemas/EnumType'
      - name: limit
        in: query
        description: "Pagination: pageSize"
        schema:
          format: int32
          default: 20
          type: integer
      - name: offset
        in: query
        description: "Pagination: pageIndex"
        schema:
          format: int32
          default: 0
          type: integer
      - name: sortDirection
        in: query
        description: Sortierung. Es wird nach SCHLUESSEL sortiert.
        schema:
          allOf:
          - $ref: '#/components/schemas/SortDirection'
          - default: asc
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RaetselsucheTreffer'
        "400":
          description: fehlgeschlagene Input-Validierung
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      security:
      - SecurityScheme:
        - ADMIN
        - AUTOR
        - LEHRER
        - PRIVAT
        - STANDARD
      - SecurityScheme:
        - ADMIN
        - AUTOR
        - LEHRER
        - PRIVAT
        - STANDARD
...
components:
  schemas:
...
    Deskriptor:
      description: Deskriptor-Entity mit allen Attributen.
      type: object
      properties:
        id:
          format: int64
          description: technische ID
          type: integer
        name:
          description: Name zum Anzeigen
          type: string
        admin:
          description: "Flag, ob dieser Deskriptor nur in der Admin-Anwendung zur\
            \ Verfügung steht"
          type: boolean
        kontext:
          description: Kontext des Deskriptors
          type: string
          deprecated: true
...
    DomainEntityStatus:
      description: Veröffentlichungsstatus eines Domain-Objekts
      enum:
      - ERFASST
      - FREIGEGEBEN
      type: string
    EditRaetselPayload:
      description: Payload zum Anlegen und Ändern eines Rätsels
      type: object
      properties:
        latexHistorisieren:
          description: "Flag, ob Änderungen im LaTeX-Code historisiert werden sollen.\
            \ Ist nur bei inhaltlichen Änderungen sinnvoll."
          type: boolean
        raetsel:
          description: die zu speichernden Daten des Rätsels
          type: object
          allOf:
          - $ref: '#/components/schemas/Raetsel'
...
    EnumType:
      enum:
      - ORDINAL
      - STRING
      type: string
...
    GeneratedFile:
      type: object
      properties:
        fileName:
          type: string
        fileData:
          format: binary
          type: string
...
    Images:
      description: Objekt das ein png mit dem Text für die Frage/Quizaufgabe und ein
        png mit dem Text für die Lösung enthält
      type: object
      properties:
        imageFrage:
          format: binary
          description: Base64-encodetes png mit dem Text der Frage/Quizaufgabe. Kann
            mit einem img src="data:image/png;base64- Tag angezeigt werden
          type: string
        imageLoesung:
          format: binary
          description: Base64-encodetes png mit dem Text der Lösung. Kann mit einem
            img src="data:image/png;base64- Tag angezeigt werden
          type: string
    LayoutAntwortvorschlaege:
      description: "enum für das gewünschte Layout der Antwortvorschläge im Aufgabenimage.\
        \ NOOP: keine Antwortvorschläge, ANKREUZTABELLE: zweizeilige Tabelle wie im\
        \ Minikänguruwettbewerb - oben die Lösungsvorschläge - unten leer zum Ankreuzen,\
        \ BUCHSTABEN: zweizeilige Tabelle - oben Lösungsbuchstaben, unten die Lösungsvorschlä\
        ge, DESCRIPTION: zweispaltige Liste - links Lösungsbuchstaben, rechts die\
        \ Lösungsvorschläge"
      enum:
      - NOOP
      - ANKREUZTABELLE
      - BUCHSTABEN
      - DESCRIPTION
      type: string
    MessagePayload:
      description: eine kategorisierte Meldung
      type: object
      properties:
        level:
          description: "Level der Message: INFO | WARN | ERROR"
          type: string
        message:
          description: die message
          type: string
...
    RaetselsucheTreffer:
      description: Teilmenge der Treffer einer Suchanfrage
      type: object
      properties:
        trefferGesamt:
          format: int64
          description: Gesamtzahl aller Treffer der Suchanfrage
          type: integer
        treffer:
          description: die abgefragte Teilmenge aller Treffer der Suchanfrage (Page)
          type: array
          items:
            type: object
            properties:
              id:
                type: string
              schluessel:
                type: string
              name:
                type: string
              kommentar:
                type: string
              status:
                $ref: '#/components/schemas/DomainEntityStatus'
              deskriptoren:
                type: array
                items:
                  $ref: '#/components/schemas/Deskriptor'
...
    SortDirection:
      enum:
      - asc
      - desc
      type: string
...